id match related bug fixed while update and deleting a product

+++ services/Product.services.js
@@ -26,19 +26,33 @@ module.exports.getSpecificProductByIdService = async (productId) => {
     const product = await Product.findById(productId).select(
         'name price unit quantity description _id status'
     );
-    return product;
+    if (!product) {
+        throw new Error('No product found with that ID');
+    } else {
+        return product;
+    }
 };
 
 module.exports.updateProductByIdService = async (productId, data) => {
-    await Product.updateOne(
-        { _id: productId },
-        { $set: data },
-        {
-            runValidators: true, // this will check if the data is valid or not depending on the schema. if this option is not set, then the data will be updated without any validation. suppose, if anyone tries to update the price of a product with a string value, then it will be updated without any error. but if this option is set to true, then it will throw an error.
-        }
-    );
+    const product = await Product.findById(productId);
+    if (!product) {
+        throw new Error('No product found with that ID');
+    } else {
+        await Product.updateOne(
+            { _id: productId },
+            { $set: data },
+            {
+                runValidators: true, // this will check if the data is valid or not depending on the schema. if this option is not set, then the data will be updated without any validation. suppose, if anyone tries to update the price of a product with a string value, then it will be updated without any error. but if this option is set to true, then it will throw an error.
+            }
+        );
+    }
 };
 
 module.exports.deleteProductByIdService = async (productId) => {
-    await Product.deleteOne({ _id: productId });
+    const product = await Product.findById(productId);
+    if (!product) {
+        throw new Error('No product found with that ID');
+    } else {
+        await Product.deleteOne({ _id: productId });
+    }
 };
